creating useMovies hooks
