+++ src/components/App.jsx
@@ -1,4 +1,4 @@
-import React, { Fragment, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import Nav from "./Nav";
 import Box from "./Box";
 import WatchedSummary from "./WatchedSummary";
@@ -9,31 +9,7 @@ import Search from "./Search";
 import MovieDetails from "./MovieDetails";
 import Loader from "./Loader";
 import ErrorMessage from "./ErrorMessage";
-
-const tempWatchedData = [
-  {
-    imdbID: "tt1375666",
-    Title: "Inception",
-    Year: "2010",
-    Poster:
-      "https://m.media-amazon.com/images/M/MV5BMjAxMzY3NjcxNF5BMl5BanBnXkFtZTcwNTI5OTM0Mw@@._V1_SX300.jpg",
-    runtime: 148,
-    imdbRating: 8.8,
-    userRating: 10,
-  },
-  {
-    imdbID: "tt0088763",
-    Title: "Back to the Future",
-    Year: "1985",
-    Poster:
-      "https://m.media-amazon.com/images/M/MV5BZmU0M2Y1OGUtZjIxNi00ZjBkLTg1MjgtOWIyNThiZWIwYjRiXkEyXkFqcGdeQXVyMTQxNzMzNDI@._V1_SX300.jpg",
-    runtime: 116,
-    imdbRating: 8.5,
-    userRating: 9,
-  },
-];
-
-const KEY = import.meta.env.VITE_APP_OMDB_API_KEY;
+import { useMovies } from "../hooks/useMovies";
 
 function App() {
   const [query, setQuery] = useState("");
@@ -41,10 +17,9 @@ function App() {
     const storedData = localStorage.getItem("watched");
     return JSON.parse(storedData);
   });
-  const [movies, setMovies] = useState([]);
   const [selectedId, setSelectedId] = useState(null);
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState("");
+
+  const { movies, isLoading, error } = useMovies(query, handleCloseMovie);
 
   function handleSelectMovie(id) {
     setSelectedId((selectedId) => (selectedId === id ? null : id));
@@ -67,49 +42,6 @@ function App() {
       localStorage.setItem("watched", JSON.stringify(watched));
     },
     [watched]
-  );
-
-  useEffect(
-    function () {
-      const controller = new AbortController();
-
-      async function fetchMovieData() {
-        try {
-          setIsLoading(true);
-          setError("");
-
-          const response = await fetch(
-            `http://www.omdbapi.com/?apikey=${KEY}&s=${query}`,
-            {
-              signal: controller.signal,
-            }
-          );
-
-          if (!response.ok) throw new Error("Failed to fetched data");
-
-          const data = await response.json();
-          if (data.Response === "False") throw new Error("Movie not found!");
-
-          setMovies(data.Search);
-          setError("");
-        } catch (err) {
-          if (err.name !== "AbortError") {
-            console.log(err.message);
-            setError(err.message);
-          }
-        } finally {
-          setIsLoading(false);
-        }
-      }
-
-      handleCloseMovie();
-      if (query.length !== 0) fetchMovieData();
-
-      return function () {
-        controller.abort();
-      };
-    },
-    [query]
   );
 
   return (
