+++ src/components/MovieDetails.jsx
@@ -1,15 +1,19 @@
-import { useState, useEffect } from "react";
+import { useState, useEffect, useRef } from "react";
 import StarRating from "./StarRating";
 import Loader from "./Loader";
 import ErrorMessage from "./ErrorMessage";
 
 const KEY = import.meta.env.VITE_APP_OMDB_API_KEY;
+const fallbackImg =
+  "https://images.unsplash.com/reserve/LJIZlzHgQ7WPSh5KVTCB_Typewriter.jpg?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxzZWFyY2h8MTJ8fG1vdmllJTIwcG9zdGVyfGVufDB8fDB8fHww&auto=format&fit=crop&w=500&q=60";
 
 function MovieDetails({ watched, selectedId, onCloseMovie, onAddWatched }) {
   const [movie, setMovie] = useState({});
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState("");
   const [userRating, setUserRating] = useState("");
+
+  const countRef = useRef([]);
 
   // const isWatched = watched.some((movie) => movie.imdbID === selectedId);
   const watchedUserRating = watched.find(
@@ -38,11 +42,20 @@ function MovieDetails({ watched, selectedId, onCloseMovie, onAddWatched }) {
       year,
       imdbRating: Number(imdbRating),
       userRating,
+      countRattingDecision: countRef.current,
     };
 
     onAddWatched(newWatchedMovie);
     onCloseMovie();
   }
+
+  useEffect(
+    function () {
+      if (!userRating) return;
+      countRef.current = [...countRef.current, userRating];
+    },
+    [userRating]
+  );
 
   useEffect(
     function () {
@@ -111,7 +124,10 @@ function MovieDetails({ watched, selectedId, onCloseMovie, onAddWatched }) {
             <button className="btn-back" onClick={() => onCloseMovie()}>
               &larr;
             </button>
-            <img src={poster} alt={`Poster of ${title} movie`} />
+            <img
+              src={poster !== "N/A" ? poster : fallbackImg}
+              alt={`Poster of ${title} movie`}
+            />
             <div className="details-overview">
               <h2>{title}</h2>
               <p>
