+++ src/components/Search.jsx
@@ -1,11 +1,34 @@
+import { useRef, useEffect } from "react";
+
 export default function Search({ query, setQuery }) {
+  const inputEl = useRef(null);
+
+  useEffect(
+    function () {
+      function callback(e) {
+        if (e.code === "Enter") {
+          if (document.activeElement === inputEl.current) return;
+
+          inputEl.current.focus();
+          setQuery("");
+        }
+      }
+      inputEl.current.focus();
+
+      document.addEventListener("keydown", callback);
+      return () => document.removeEventListener("keydown", callback);
+    },
+    [setQuery]
+  );
+
   return (
     <input
+      className="search"
       type="text"
       placeholder="Search movies..."
       value={query}
       onChange={(e) => setQuery(e.target.value)}
-      className="search"
+      ref={inputEl}
     />
   );
 }
