+++ src/hooks/useMovies.js
@@ -1,0 +1,54 @@
+import { useState, useEffect } from "react";
+
+const KEY = import.meta.env.VITE_APP_OMDB_API_KEY;
+
+export function useMovies(query, callback) {
+  const [movies, setMovies] = useState([]);
+  const [isLoading, setIsLoading] = useState(false);
+  const [error, setError] = useState("");
+
+  useEffect(
+    function () {
+      // callback?.();
+      const controller = new AbortController();
+
+      async function fetchMovieData() {
+        try {
+          setIsLoading(true);
+          setError("");
+
+          const response = await fetch(
+            `http://www.omdbapi.com/?apikey=${KEY}&s=${query}`,
+            {
+              signal: controller.signal,
+            }
+          );
+
+          if (!response.ok) throw new Error("Failed to fetched data");
+
+          const data = await response.json();
+          if (data.Response === "False") throw new Error("Movie not found!");
+
+          setMovies(data.Search);
+          setError("");
+        } catch (err) {
+          if (err.name !== "AbortError") {
+            console.log(err.message);
+            setError(err.message);
+          }
+        } finally {
+          setIsLoading(false);
+        }
+      }
+
+      if (query.length !== 0) fetchMovieData();
+
+      return function () {
+        controller.abort();
+      };
+    },
+    [query]
+  );
+
+  return { movies, isLoading, error };
+}
